requirements.txt
python-gitlab==1.0.0
xlwt==1.3.0
click==7.0
pyinstaller